상품 매출 추가 (+ 코드 수정)

+++ src/main/java/project1/shop/adminService/AdminService.java
@@ -94,6 +94,10 @@ public class AdminService {
     @Transactional
     public void deleteAdmin(Long adminId) {
 
+        if (adminId == 1){
+            throw new IllegalArgumentException("대표의 아이디는 삭제가 불가능합니다.");
+        }
+
         Admin admin = adminRepository.findById(adminId).orElseThrow(IllegalArgumentException::new);
 
         admin.delete();
